Resources refactoring, game variable checks, fixes
+ Changed some of the resources to use chunk id instead of index number.
+ Buttons, doors and recharge station now check game variables for
access permission.
+ Started music importing.
+ Fixed material animations
+ Unity has changed font rendering again. Fixed.
+ Fixed MessageBus

+++ SSImporter/Editor/Inspector/StringLibraryInspector.cs
@@ -13,14 +13,14 @@ public class StringLibraryInspector : InspectorBase<StringLibrary> {
         EditorGUIUtility.labelWidth = 50;
 
         CyberString[] Strings = Target.Strings;
-        uint[] ChunkIds = Target.ChunkIds;
+        ushort[] ChunkIds = Target.ChunkIds;
 
         if (foldOuts == null)
             foldOuts = new bool[ChunkIds.Length];
 
         for (int chunkIndex = 0; chunkIndex < ChunkIds.Length; ++chunkIndex) {
             string[] chunkStrings = Strings[chunkIndex].Strings;
-            uint chunkId = ChunkIds[chunkIndex];
+            ushort chunkId = ChunkIds[chunkIndex];
 
             if (foldOuts[chunkIndex] = EditorGUILayout.Foldout(foldOuts[chunkIndex], chunkId.ToString())) {
                 ++EditorGUI.indentLevel;
